+++ src/renderer/src/sub-components/KatexRenderer.vue
