+++ src/renderer/src/sub-components/SvgIcon.vue
