+++ src/renderer/src/sub-components/SymbolsItem.vue
